Reorganize Slate helpers
To distinguish between functions that mutate the Change object and
smaller utility functions.
