+++ client/src/slate-helpers/slate-change-mutations/adjustCursorIfAtEdge.ts
@@ -1,5 +1,5 @@
 import { Change } from "../../components/BlockEditor/types";
-import { isCursorInPotentiallyProblematicPosition } from "./isCursorInPotentiallyProblematicPosition";
+import { isCursorInPotentiallyProblematicPosition } from "../slate-utils/isCursorInPotentiallyProblematicPosition";
 export function adjustCursorIfAtEdge(change: Change) {
   const value = change.value;
   const selectionIsExpanded = value.selection.isExpanded;
