+++ client/src/slate-helpers/slate-change-mutations/handleCursorNavigationAcrossPointerEdge.ts
