+++ client/src/slate-helpers/slate-change-mutations/moveSelectionAwayFromPointerEdge.ts
