+++ client/src/slate-helpers/slate-change-mutations/normalizeAfterRemoval.ts
