+++ client/src/slate-helpers/slate-utils/closestPointerExportAncestor.ts
