+++ client/src/slate-helpers/slate-utils/hasSpacerAsFirstChar.ts
