+++ client/src/slate-helpers/slate-utils/hasSpacerAsLastChar.ts
