+++ client/src/slate-helpers/slate-utils/isCursorInPotentiallyProblematicPosition.ts
