+++ client/src/slate-helpers/slate-utils/isFirstTextOfExport.ts
