+++ client/src/slate-helpers/slate-utils/isLastTextOfExport.ts
